+++ src/workspaces/ShareWorkspaceModal/ShareWorkspaceModal.test.ts
@@ -5,10 +5,10 @@ import { h } from 'react-hyperscript-helpers';
 import { Ajax } from 'src/libs/ajax';
 import { getTerraUser } from 'src/libs/state';
 import { GoogleWorkspace } from 'src/libs/workspace-utils';
-import { AccessEntry, RawWorkspaceAcl } from 'src/pages/workspaces/workspace/WorkspaceAcl';
 import { asMockedFn, renderWithAppContexts as render } from 'src/testing/test-utils';
 import { defaultAzureWorkspace, defaultGoogleWorkspace, protectedAzureWorkspace } from 'src/testing/workspace-fixtures';
 import ShareWorkspaceModal from 'src/workspaces/ShareWorkspaceModal/ShareWorkspaceModal';
+import { AccessEntry, RawWorkspaceAcl } from 'src/workspaces/WorkspaceAcl';
 
 jest.mock('src/libs/state', () => ({
   ...jest.requireActual('src/libs/state'),
