+++ src/workspaces/hooks/useDeleteWorkspaceState.test.ts
@@ -4,8 +4,8 @@ import { generateTestApp } from 'src/analysis/_testData/testData';
 import { Ajax } from 'src/libs/ajax';
 import { reportError } from 'src/libs/error';
 import { AzureWorkspaceInfo, BaseWorkspace, GoogleWorkspaceInfo } from 'src/libs/workspace-utils';
-import { useDeleteWorkspaceState } from 'src/pages/workspaces/hooks/useDeleteWorkspaceState';
 import { asMockedFn, renderHookInAct } from 'src/testing/test-utils';
+import { useDeleteWorkspaceState } from 'src/workspaces/hooks/useDeleteWorkspaceState';
 
 type AjaxExports = typeof import('src/libs/ajax');
 jest.mock('src/libs/ajax', (): AjaxExports => {
